Add files via upload
Script for archiving and sending files from a source destination to a destination backup server. The script will first display the size of the source directory, archive the files that is older than 10 days and send it to a backup server. If the backup was successful, the original files will be deleted to free up disk space